get step buttons helper implemented
